ENGR00230369 imx6 sabresd: refine the keyguard lock screen layout
When the system font size is set to quite large,
the digital time widget can not show the full time
as the keyguard screen status land layout width
is too small. Here just decrease the layout left and
right margin to display the time correctly.

The actual differences is like:
@@ -42,9 +42,9 @@
<include layout="@layout/keyguard_screen_status_land"
		 android:layout_width="match_parent"
		 android:layout_height="wrap_content"
-                android:layout_marginLeft="50dip"
+                android:layout_marginLeft="20dip"
		 android:layout_marginTop="50dip"
-                android:layout_marginRight="64dip"
+                android:layout_marginRight="30dip"
		 android:layout_alignParentTop="true"
		 android:layout_alignParentLeft="true"/>

Change-Id: I39c9c1f8dcfc4aee331a60308ba55efc3c0f770a
Signed-off-by: Xinyu Chen <xinyu.chen@freescale.com>